Refactoring the layout component
