+++ src/components/docs/DirectoryOverflow.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Bars3Icon } from "@heroicons/react/24/outline";
 import { useEffect, useRef, useState } from "react";
 import { DocsNavigationList } from "./DocsNavigationList";
