+++ src/components/docs/LeftHandSideParent.tsx
@@ -2,7 +2,7 @@
 
 import { DocsNavigationList } from "./DocsNavigationList";
 
-export const LeftHandSideParentContainer = ({
+export const LeftHandSideBarContainer = ({
   tableOfContents,
   header,
 }: {
@@ -12,7 +12,6 @@ export const LeftHandSideParentContainer = ({
   return (
     <div className="h-5/6 w-full rounded-2xl bg-white/50 shadow-xl">
       {header}
-      {/* eslint-disable-next-line tailwindcss/no-arbitrary-value */}
       <div className="h-[76%] overflow-y-auto overflow-x-hidden pl-4 2xl:max-h-[75vh] 2xl:pl-0">
         <DocsNavigationList navItems={tableOfContents} />
       </div>
